kvutils: Transaction validation refactoring [KVL-1015] CHANGELOG_BEGIN CHANGELOG_END
