+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/committer/transaction/validation/TransactionValidator.scala
@@ -1,0 +1,10 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.kvutils.committer.transaction.validation
+
+import com.daml.ledger.participant.state.kvutils.committer.transaction.{Step, TransactionRejector}
+
+private[transaction] trait TransactionValidator {
+  def createValidationStep(rejector: TransactionRejector): Step
+}
