+++ ledger/participant-state/kvutils/src/test/lib/scala/com/daml/ledger/participant/state/kvutils/TestHelpers.scala
@@ -7,12 +7,22 @@ import java.time.Duration
 import java.util.UUID
 
 import com.daml.daml_lf_dev.DamlLf
-import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlLogEntryId
-import com.daml.ledger.participant.state.kvutils.committer.CommitContext
+import com.daml.ledger.participant.state.kvutils.DamlKvutils.{
+  DamlLogEntryId,
+  DamlSubmitterInfo,
+  DamlTransactionEntry,
+  DamlTransactionRejectionEntry,
+}
+import com.daml.ledger.participant.state.kvutils.committer.transaction.DamlTransactionEntrySummary
+import com.daml.ledger.participant.state.kvutils.committer.{CommitContext, StepResult, StepStop}
 import com.daml.ledger.participant.state.v1.{Configuration, ParticipantId, TimeModel}
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Time.Timestamp
+import com.daml.lf.transaction.SubmittedTransaction
+import com.daml.lf.transaction.test.TransactionBuilder
 import com.google.protobuf.ByteString
+
+import scala.jdk.CollectionConverters._
 
 object TestHelpers {
   def name(value: String): Ref.Name = Ref.Name.assertFromString(value)
@@ -48,4 +58,29 @@ object TestHelpers {
       participantId: Int = 0,
   ): CommitContext =
     CommitContext(inputs, recordTime, mkParticipantId(participantId))
+
+  def createEmptyTransactionEntry(submitters: List[String]): DamlTransactionEntry =
+    createTransactionEntry(submitters, TransactionBuilder.EmptySubmitted)
+
+  def createTransactionEntry(
+      submitters: List[String],
+      tx: SubmittedTransaction,
+  ): DamlTransactionEntry =
+    DamlTransactionEntry.newBuilder
+      .setTransaction(Conversions.encodeTransaction(tx))
+      .setSubmitterInfo(
+        DamlSubmitterInfo.newBuilder
+          .setCommandId("commandId")
+          .addAllSubmitters(submitters.asJava)
+      )
+      .setSubmissionSeed(ByteString.copyFromUtf8("a" * 32))
+      .build
+
+  def getTransactionRejectionReason(
+      result: StepResult[DamlTransactionEntrySummary]
+  ): DamlTransactionRejectionEntry =
+    result
+      .asInstanceOf[StepStop]
+      .logEntry
+      .getTransactionRejectionEntry
 }
